services/grounding_service.py
"""Google Search Grounding service."""

from typing import Dict, List, Optional
from loguru import logger

try:
    from google.cloud import aiplatform
    from google.cloud.aiplatform.gapic.schema import predict
    GROUNDING_AVAILABLE = True
except ImportError:
    GROUNDING_AVAILABLE = False
    logger.warning("Google Search Grounding not available")

from config.settings import settings


class GroundingService:
    """Service for Google Search Grounding."""
    
    def __init__(self):
        """Initialize grounding service."""
        self.available = GROUNDING_AVAILABLE and settings.google_cloud_project_id is not None
        
        if self.available:
            try:
                aiplatform.init(
                    project=settings.google_cloud_project_id,
                    location=settings.google_cloud_region
                )
                logger.info("Grounding service initialized")
            except Exception as e:
                logger.error(f"Error initializing grounding service: {str(e)}")
                self.available = False
        else:
            logger.info("Grounding service not available, using fallback")
    
    def retrieve_academic_context(self, query: str) -> Dict:
        """
        Retrieve academic context using Google Search Grounding.
        
        Args:
            query: Search query (e.g., "Western Blot troubleshooting")
            
        Returns:
            Dictionary with citations and context
        """
        if not self.available:
            # Fallback: return mock data
            return {
                "citations": [
                    {
                        "title": "Common Western Blot Issues",
                        "url": "https://example.com/western-blot-troubleshooting",
                        "snippet": "Western Blot troubleshooting guide..."
                    }
                ],
                "context": f"Academic literature suggests that {query} is a common issue in biochemistry research.",
                "hashtags": ["#Research", "#Biochemistry", "#LabLife"]
            }
        
        # Vertex AI Grounding implementation would go here
        # This is a placeholder for the full implementation
        try:
            # Placeholder for actual grounding API call
            return {
                "citations": [],
                "context": "",
                "hashtags": []
            }
        except Exception as e:
            logger.error(f"Error in grounding: {str(e)}")
            return {
                "citations": [],
                "context": "",
                "hashtags": []
            }

