requirements.txt
# Core AI Services
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Agent Framework (optional - for future LangChain integration)
langchain>=0.1.0
langchain-google-genai>=0.0.5

# Frontend (deprecated - using React/Next.js now)
# streamlit>=1.28.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Handling
numpy>=1.24.0
pandas>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Vector Search (for local alternative)
faiss-cpu>=1.7.4

# Clustering & Visualization
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
tenacity>=8.2.0

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

